CodeGen from PR 19983 in Azure/azure-rest-api-specs Merge 36f190aa6f5cd4485bb3ec1f9632a3e7dcc6b9e3 into 64d1545fcbb9c6b4f41e2fd1fd2f3e7cb421772d

+++ sdk/resourcemanager/iotcentral/armiotcentral/models.go
@@ -136,7 +136,7 @@ type AppTemplate struct {
 	Name *string `json:"name,omitempty" azure:"ro"`
 
 	// READ-ONLY; The order of the template in the templates list.
-	Order *float32 `json:"order,omitempty" azure:"ro"`
+	Order *int32 `json:"order,omitempty" azure:"ro"`
 
 	// READ-ONLY; The title of the template.
 	Title *string `json:"title,omitempty" azure:"ro"`
